bot/requirements.txt
aiogram>=3.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
